Added update option which supports adding additional, new images from another source directory (the images are tracked witha SHA256 checksum in data.json so duplicates are ignored).
Update is implicit if the destination directory looks like a gallery,
i.e. it contains a data.json file and a view subdirectory.

Removal of images from the img subdirectory is also supported, in
which case the -u option is required when operating on the one
gallery directory.

Update rewrites data.json as well as index.html and thus can also
be used to simply replace titles/names in the html files.  It also
recreates the view subdirectory so existing galleries can get the
new JavaScript and CSS files if they didn't use symlinks to the
shared view directory.