Add files in batches to the archive
This should avoid a crash when large number of files are processed.
Note that we could query `getconf MAX_ARG` and build our command line
accordingly. But what would we do on Windows? So now we're using
batches of 200 images each. The code is based on this example:

use Modern::Perl;
my @f = 1 .. 100;
my $n = 30;
for my $i (0 .. @f/$n) {
  my $m = $n;
  $m = @f % $n if $i >= int(@f/$n);
  last unless $m;
  my @g = @f[$i*$n .. $i*$n + $m - 1];
  say "@g";
}

The change does not extend to deleting files. We're assuming we don't
want to delete hundred of files.

Note that fgallery has some code "make paths explicitly
absolute/relative to strip file path info with 7za". We seem to be
lacking this code.
https://github.com/wavexx/fgallery/blob/master/fgallery#L801